shift the inclusion of <grackle.h> from _enzo.hpp to chemistry.hpp (in the future, I plan to totally remove it from public headers in the future
